+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - chat: Improve sorting in modes report and highlight active instead of using a `*`
 - chat: Remove sessions report active column and highlight active id
 - chat: Show `0` messages in the sessions report with a dark gray color
+- chat: The /session-delete command now supports `all` as an id to remove all sessions
 
 ### Internal
 - internal: Rename reportings plain() to style() for clarity